feat(RSSFeedReaderClient): support for atom rss feeds
Not tested. I went through this client in order to write tests for it. I realized that the client did not use metrics, could have better error handling and support for atom rss feeds. I can add another label to the success metric (possibly the fail metric as well) called the rss type (rss or atom).