Refactor PrometheusController + Tests + Errors, MetricsService and MessageService and their Dependencies to AutomaUtilities (#198)
* fix: added comments for what must be refactored, small typo fix

* refactor: prometheus controller, message service and metrics service to
automa utilities package

Need to merge automa utilities refactor branch to master, update packages in this repository, and build package and fix any small errors that might arise

* fixup! refactor: prometheus controller, message service and metrics service to automa utilities package

* refactor(BackendMetric): removed message service metric

MessageService is in AutomaUtilities package now. Metric removed in this commit is added in AutomaUtilities.

* feat(configure): use prometheus service instead of legacy prometheus controller from automa utilities

This removes the need of an authentication token to prevent leaking metrics. It works, because `PrometheusService` starts up a separate HTTP server on a different port that is not exposed publicly.

* fix(docker-compose): pass github_ssh_authentication_token to api_dev so that automa utilities package can be cloned

* fix(package.json): set github token build secret when deploying to fly

This is needed to clone AutomaUtilities, a private Github repository. This commit also exposes the metrics port in both dockerfiles

* fix(BackendMetric): use metric status data type from automa utilities

* feat(AutomaWebCoreClient): created client to interact with automa web core api

Made method to get website html. Updated AutomaUtilities to have the web core api endpoint payload type.

* fix(automa-backend-testing.yml): set gh_pat so that swift test opensource action can clone private repo - automautilities

* fixup! fix(automa-backend-testing.yml): set gh_pat so that swift test opensource action can clone private repo - automautilities